+++ components/restreamer/client/src/components/AppDashboard.svelte
@@ -1,6 +1,7 @@
 <script lang="js">
-  import { createGraphQlClient } from '../utils/util';
-  import { setClient, subscribe } from 'svelte-apollo';
+  import Confirm from './common/Confirm.svelte';
+  import { createGraphQlClient, showError } from '../utils/util';
+  import { mutation, setClient, subscribe } from 'svelte-apollo';
   import Shell from './common/Shell.svelte';
   import { Statistics } from '../../api/dashboard.graphql';
   import ToolbarDashboard from './ToolbarDashboard.svelte';
@@ -8,6 +9,10 @@
   import StatusFilter from './common/StatusFilter.svelte';
   import { statusesList } from '../utils/constants';
   import { toggleFilterStatus } from '../utils/filters.util';
+  import {
+    EnableAllOutputsForClients,
+    DisableAllOutputsForClients,
+  } from '../../api/dashboard.graphql';
 
   const gqlClient = createGraphQlClient(
     '/api-dashboard',
@@ -18,6 +23,12 @@
 
   let isOnline = false;
   const dashboard = subscribe(Statistics, { errorPolicy: 'all' });
+  const enableAllOutputsForClientMutation = mutation(
+    EnableAllOutputsForClients
+  );
+  const disableAllOutputsForClientMutation = mutation(
+    DisableAllOutputsForClients
+  );
 
   let title = document.title;
   $: document.title = (isOnline ? '' : '🔴  ') + title;
@@ -89,6 +100,22 @@
         )
       : [];
   }
+
+  async function enableAllOutputsOfAllRestreams() {
+    try {
+      await enableAllOutputsForClientMutation();
+    } catch (e) {
+      showError(e.message);
+    }
+  }
+
+  async function disableAllOutputsOfAllRestreams() {
+    try {
+      await disableAllOutputsForClientMutation();
+    } catch (e) {
+      showError(e.message);
+    }
+  }
 </script>
 
 <template>
@@ -113,7 +140,7 @@
               {/each}
             </span>
           </div>
-          <div class="uk-width-1-4@m">
+          <div class="uk-flex-auto uk-flex-right uk-flex uk-flex-middle">
             <span class="toolbar-label">
               OUTPUTS:
 
@@ -127,6 +154,37 @@
                 />
               {/each}
             </span>
+          </div>
+          <div class="uk-margin-auto-left">
+            <Confirm let:confirm>
+              <button
+                class="uk-button uk-button-default"
+                data-testid="start-all-outputs"
+                title="Start all outputs of all restreams"
+                on:click={() => confirm(enableAllOutputsOfAllRestreams)}
+                ><span>Start All</span>
+              </button>
+              <span slot="title">Start all outputs</span>
+              <span slot="description"
+                >This will start all outputs of all restreams.
+              </span>
+              <span slot="confirm">Start</span>
+            </Confirm>
+
+            <Confirm let:confirm>
+              <button
+                class="uk-button uk-button-default"
+                data-testid="stop-all-outputs"
+                title="Stop all outputs of all restreams"
+                on:click={() => confirm(disableAllOutputsOfAllRestreams)}
+                ><span>Stop All</span>
+              </button>
+              <span slot="title">Stop all outputs</span>
+              <span slot="description"
+                >This will stop all outputs of all restreams.
+              </span>
+              <span slot="confirm">Stop</span>
+            </Confirm>
           </div>
         </div>
       </section>